Moved all the services to designeted folder
