+++ src/app/app-module.ts
@@ -6,14 +6,14 @@ import { App } from './app';
 import { Header } from './header/header';
 import { DynamicForm } from './auth/dynamic-form/dynamic-form';
 import { Login } from './auth/login';
-import { USERS_API_SERVICE_TOKEN, usersServiceFactory } from './UsersApiService';
+import { USERS_API_SERVICE_TOKEN, usersServiceFactory } from './services/usersApi/UsersApiService';
 import { AkitaNgDevtools } from '@datorama/akita-ngdevtools';
 import { Register } from './auth/register';
 import { Main } from './main/main';
 import { ShoeCard } from './components/shoe-card/shoe-card';
 import { LogoHeadline } from './components/logo-headline/logo-headline';
 import { persistState } from '@datorama/akita';
-import { shoesApiService } from './shoesApiService';
+import { shoesApiService } from './services/shoesApi/shoesApiService';
 import { TopPicks } from './components/top-picks/top-picks';
 import { Shop } from './shop/shop';
 import { ShopSidebar } from './shop/shop-sidebar/shop-sidebar';
